Update README for presentation, modify gpucheck
The README now provides a roadmap for parsing the large number of files
included in this repo. It also suggests a comprehensible path forward
for beginning to run these files.
gpucheck will now print out a corresponding message if no CUDA device is
found. Next steps are to print out info for each CUDA device available.
Additionally, making a header file to deal with standard matrix
techniques is probably a must for readability of the files.